Fix odd behavior with `base_level` method
The `base_level` method is actually removed. The base level, or the
level when the verbosity is zero (0), is now inherited from the
environment via the `RUST_LOG` variable. This makes the `base_level`
method obsolete, but also difficult to reconcile with the value set for
the environment versus when set in the code.